+++ src/Blogifier/Pages/FakeEditor.razor
@@ -4,6 +4,7 @@
 @using System.Net.Http
 @using Microsoft.AspNetCore.Http
 
+@inject IJSRuntime JSRuntime
 @inject IHttpClientFactory ClientFactory
 @inject IStringLocalizer<Resource> _localizer
 
@@ -18,18 +19,18 @@
                     d="M0 8a8 8 0 1 1 16 0A8 8 0 0 1 0 8zm8-7a7 7 0 0 0-5.468 11.37C3.242 11.226 4.805 10 8 10s4.757 1.225 5.468 2.37A7 7 0 0 0 8 1z" />
             </svg>
         </div>
-        <div class="col my-auto my-0 py-0" style="height: 100px;">
-            <form method="post" action="/Home/ExternalSignIn" class="d-grid mx-auto my-0 h-100">
+        <div class="col my-auto py-auto d-grid" style="height: 150px;">
+            @* <form method="post" action="/Home/ExternalSignIn" class="d-grid mx-auto my-0 h-100">
                 <div class="form-group">
-                    <input class="form-control" name="path" hidden="true" aria-label="currentpath"
-                        value="@CurrentPath" />
+                    <input class="form-control" name="path" hidden="true" aria-label="currentpath" id="hidden_returnpath" />
                 </div>
                 <button class="btn btn-light h-100" type="submit" aria-label="@_localizer["login"]">
                     Register or Login to Comment!
                 </button>
                 <div class="form-group">
-                </div>
-            </form>
+                </div> 
+            </form> *@
+            <a class="d-grid btn btn-light h-50 my-auto py-auto align-items-center" href="@ReturnPath" role="button">Register or Login to Comment!</a>
         </div>
     </div>
     @* <div class="row">
@@ -46,18 +47,19 @@
 
 @code {
     [Parameter] public string CurrentPath { get; set; }
-    /*private HttpClient _httpClient { get; set; }
-    private SubmitModel submitModel { get; set; }
+    private string ReturnPath {get; set;}
     protected override async Task OnInitializedAsync()
     {
     await Load();
     }
     public async Task Load()
     {
-    // _httpClient = ClientFactory.CreateClient("Local");
-    // await Task.Run(() => submitModel = new SubmitModel() { ReturnPath = CurrentPath });
+        await Task.Run(() =>ReturnPath = "Oidc/ExternalSignIn/?path=" + CurrentPath);
     }
-    private void HandleValidSubmit()
+    
+    /*protected override async Task OnAfterRenderAsync(bool firstRender)
     {
+        if (firstRender){}
+        await JSRuntime.InvokeAsync<string>("commonJsFunctions.setOidcForm", CurrentPath);
     }*/
 }
