+++ src/Blogifier/Views/Themes/standard/components/nav.cshtml
@@ -16,6 +16,7 @@
     var currentUserAuthor = await _authorProvider.FindByOpenId(User.FindFirstValue(JwtClaimTypes.Subject));
 
     var currentPath = this.Context.Request.Path;
+    var returnPath = "Oidc/ExternalSignIn/?path=" + currentPath.ToString();
     if (User.Identity.IsAuthenticated && currentUserAuthor is null)
     {
         string tempName = User.FindFirstValue(JwtClaimTypes.Name);
@@ -178,22 +179,34 @@
     }
     else
     {
-        <form method="post" asp-controller="Home" asp-action="ExternalSignIn">
+        @* <form method="post" asp-controller="Home" asp-action="ExternalSignIn">
             <div class="form-group">
-                <input class="form-control" name="currentpath" hidden="true" value="@currentPath" />
+            <input class="form-control" name="currentpath" hidden="true" value="@currentPath" />
             </div>
             <button class="header-nav-button -login ms-4" type="submit" title="@_localizer["login"]"
             aria-label="@_localizer["login"]" role="button">
 
-                <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor"
-                class="bi bi-box-arrow-in-right" viewBox="0 0 16 16">
-                    <path fill-rule="evenodd"
-                    d="M6 3.5a.5.5 0 0 1 .5-.5h8a.5.5 0 0 1 .5.5v9a.5.5 0 0 1-.5.5h-8a.5.5 0 0 1-.5-.5v-2a.5.5 0 0 0-1 0v2A1.5 1.5 0 0 0 6.5 14h8a1.5 1.5 0 0 0 1.5-1.5v-9A1.5 1.5 0 0 0 14.5 2h-8A1.5 1.5 0 0 0 5 3.5v2a.5.5 0 0 0 1 0v-2z" />
-                    <path fill-rule="evenodd"
-                    d="M11.854 8.354a.5.5 0 0 0 0-.708l-3-3a.5.5 0 1 0-.708.708L10.293 7.5H1.5a.5.5 0 0 0 0 1h8.793l-2.147 2.146a.5.5 0 0 0 .708.708l3-3z" />
-                </svg>
+            <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor"
+            class="bi bi-box-arrow-in-right" viewBox="0 0 16 16">
+            <path fill-rule="evenodd"
+            d="M6 3.5a.5.5 0 0 1 .5-.5h8a.5.5 0 0 1 .5.5v9a.5.5 0 0 1-.5.5h-8a.5.5 0 0 1-.5-.5v-2a.5.5 0 0 0-1 0v2A1.5 1.5 0 0 0
+            6.5 14h8a1.5 1.5 0 0 0 1.5-1.5v-9A1.5 1.5 0 0 0 14.5 2h-8A1.5 1.5 0 0 0 5 3.5v2a.5.5 0 0 0 1 0v-2z" />
+            <path fill-rule="evenodd"
+            d="M11.854 8.354a.5.5 0 0 0 0-.708l-3-3a.5.5 0 1 0-.708.708L10.293 7.5H1.5a.5.5 0 0 0 0 1h8.793l-2.147 2.146a.5.5 0
+            0 0 .708.708l3-3z" />
+            </svg>
             </button>
-        </form>
+            </form> *@
+        <a class="header-nav-button -login ms-4" title="@_localizer["login"]" aria-label="@_localizer["login"]"
+        role="button" href="@returnPath">
+            <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor"
+            class="bi bi-box-arrow-in-right" viewBox="0 0 16 16">
+                <path fill-rule="evenodd"
+                d="M6 3.5a.5.5 0 0 1 .5-.5h8a.5.5 0 0 1 .5.5v9a.5.5 0 0 1-.5.5h-8a.5.5 0 0 1-.5-.5v-2a.5.5 0 0 0-1 0v2A1.5 1.5 0 0 0 6.5 14h8a1.5 1.5 0 0 0 1.5-1.5v-9A1.5 1.5 0 0 0 14.5 2h-8A1.5 1.5 0 0 0 5 3.5v2a.5.5 0 0 0 1 0v-2z" />
+                <path fill-rule="evenodd"
+                d="M11.854 8.354a.5.5 0 0 0 0-.708l-3-3a.5.5 0 1 0-.708.708L10.293 7.5H1.5a.5.5 0 0 0 0 1h8.793l-2.147 2.146a.5.5 0 0 0 .708.708l3-3z" />
+            </svg>
+        </a>
     }
     <!--/Admin Menu -->
