+++ src/Blogifier/Views/Themes/test/layouts/_main.cshtml
@@ -13,7 +13,7 @@
     if (Model.ToString().EndsWith("PostModel"))
     {
         PostModel postModel = (PostModel)Model;
-        pageTitle = postModel.Post.Title + " | " +  siteTitle;
+        pageTitle = postModel.Post.Title + " | " + siteTitle;
         pageDesc = postModel.Post.Description.StripHtml();
         pageCanonical = absoluteUrl + "/posts/" + postModel.Post.Slug;
     }
@@ -21,6 +21,7 @@
 
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
@@ -50,7 +51,9 @@
         }
     </style>
 
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-BmbxuPwQa2lc/FVzBcNJ7UAyJxM6wuqIj61tLrc4wSX0szH/Ev+nYRRuWlolflfl" crossorigin="anonymous">
+    @* <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-BmbxuPwQa2lc/FVzBcNJ7UAyJxM6wuqIj61tLrc4wSX0szH/Ev+nYRRuWlolflfl" crossorigin="anonymous"> *@
+    <link href="~/themes/standard/css/bootstrap.min.css" rel="stylesheet" asp-append-version="true">
     <link rel="stylesheet" href="~/themes/standard/css/styles.css" asp-append-version="true">
 
     @Html.Raw(Model.Blog.HeaderScript)
@@ -62,7 +65,10 @@
     @RenderBody()
     <partial name="../components/footer.cshtml" />
 
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/js/bootstrap.bundle.min.js" integrity="sha384-b5kHyXgcpbZJO/tY9Ul7kGkf1S0CWuKcCD38l8YkeH8z8QjE0GmW1gYU5S9FOnJ0" crossorigin="anonymous"></script>
+    @* <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-b5kHyXgcpbZJO/tY9Ul7kGkf1S0CWuKcCD38l8YkeH8z8QjE0GmW1gYU5S9FOnJ0"
+        crossorigin="anonymous"></script> *@
+    <script src="~/themes/standard/js/frameworks/bootstrap.bundle.min.js" asp-append-version="true"></script>
     <script src="~/themes/standard/js/highlight.js" asp-append-version="true"></script>
     <script src="~/themes/standard/js/scripts.js" asp-append-version="true"></script>
     <script src="~/themes/standard/js/newsletter.js" asp-append-version="true"></script>
