+++ src/Blogifier/wwwroot/themes/standard/js/editor.js
@@ -351,6 +351,11 @@ window.commonJsFunctions = {
   getEditorValue: function () {
     return easymde.value();
   },
+  // setOidcForm: function (path) {
+  //   var inputField = document.getElementById('hidden_returnpath');
+  //   inputField.setAttribute('value', path); 
+  //   console.log('JS get Form here');
+  // },
   // Add Clear Value Function
   clearEditorValue: function () {
     easymde.value('');
@@ -381,18 +386,6 @@ window.commonJsFunctions = {
     time();
     setInterval(time, 60 * 1000);
   },
-
-  //Take the text to hidden area to be uploaded
-  // getTextToHiddenField: function () {
-  //   console.log("loaded from editor");
-  //   easymde.codemirror.on("change", () =>
-  //   {
-  //     console.log(easymde.value());
-  //     let hiddenText = document.getElementById('hidden-text');
-  //     hiddenText.value = easymde.value();
-  //   });
-  // },
-
   //Alert to debug
   tempAlert: function (message) {
     alert(message);
